feature(UX): Phase 5 report user testing Closes #3374
