+++ packages/geoview-core/public/locales/en/translation.json
@@ -42,7 +42,7 @@
     "arianavbar": "Vertical button group for map navigation",
     "basemap": "Change basemap",
     "fullscreen": "Full screen",
-    "home": "Reset to initial view",
+    "home": "Zoom to initial view",
     "zoomIn": "Zoom in",
     "zoomOut": "Zoom out",
     "coordinates": "Toggle coordinates format",
@@ -129,7 +129,7 @@
     "itemsCount": "{count} of {totalCount} classes",
     "noLayersAdded": "No layers added to the map",
     "noLayersAddedDescription": "Add layers to the map by clicking on the 'Layers' button and adding the layers you want to display.",
-    "selectLayerAndScroll": "Show in Layers panel",
+    "selectLayerAndScroll": "Manage layer",
     "hideText": "Hide feature text labels",
     "showText": "Show feature text labels",
     "layerOutOfRangeScreenReader": "(visibility limited by scale)"
@@ -256,7 +256,7 @@
     "moveLayerUp": "Move layer up",
     "moveLayerDown": "Move layer down",
     "layerControls": "Layers controls",
-    "selectLayerAndScrollTimeSlider": "Show in Time Slider panel",
+    "selectLayerAndScrollTimeSlider": "Access Time Slider panel",
     "selectLayer": "Select {{layerName}}",
     "layerActions": "Layer actions",
     "settings": {
@@ -323,7 +323,7 @@
     "ntsMapsheet": "NTS Mapsheet",
     "elevation": "Elevation",
     "guideControls": "Guide controls",
-    "selectLayerAndScrollChart": "Show in Chart panel",
+    "selectLayerAndScrollChart": "Access Chart panel",
     "featureInfoTable": "Feature information",
     "featureInfoTableField": "Field",
     "featureInfoTableValue": "Value",
@@ -455,8 +455,8 @@
     "noFeatures": "unknown...",
     "showing": "showing",
     "total": "total",
-    "selectLayerAndScroll": "Show in Data Table panel",
-    "accessAdvancedFunctions": "Access Data Table panel for advanced functions",
+    "selectLayerAndScroll": "Access Data Table panel",
+    "accessAdvancedFunctions": "Access Data Table panel",
     "filterToggle": "filters",
     "searchInputLabel": "Search data table columns",
     "tableControls": "Data table controls",
