+++ packages/geoview-core/public/locales/fr/translation.json
@@ -42,7 +42,7 @@
     "arianavbar": "Groupe de buttons vertical pour navigation sur la carte",
     "basemap": "Changer la carte de base",
     "fullscreen": "Plein écran",
-    "home": "Retour à la vue initiale",
+    "home": "Zoom à la vue initiale",
     "zoomIn": "Zoom avant",
     "zoomOut": "Zoom arrière",
     "coordinates": "Basculer le format des coordonnées",
@@ -129,7 +129,7 @@
     "itemsCount": "{count} sur {totalCount} classes",
     "noLayersAdded": "Aucune couche ajoutée",
     "noLayersAddedDescription": "Ajoutez des couches à la carte en cliquant sur le bouton 'Couches' et en sélectionnant les couches que vous souhaitez afficher.",
-    "selectLayerAndScroll": "Voir dans le panneau Couches",
+    "selectLayerAndScroll": "Gérer la couche",
     "hideText": "Masquer les étiquettes de texte des éléments",
     "showText": "Afficher les étiquettes de texte des éléments",
     "layerOutOfRangeScreenReader": "(visibilité limitée par l'échelle)"
@@ -256,7 +256,7 @@
     "moveLayerUp": "Déplacer la couche vers le haut",
     "moveLayerDown": "Déplacer la couche vers le bas",
     "layerControls": "Commandes des couches",
-    "selectLayerAndScrollTimeSlider": "Voir dans le panneau Curseur Temporel",
+    "selectLayerAndScrollTimeSlider": "Accéder au panneau Curseur Temporel",
     "selectLayer": "Sélectionner {{layerName}}",
     "layerActions": "Actions sur les couches",
     "settings": {
@@ -323,7 +323,7 @@
     "ntsMapsheet": "Carte SNRC",
     "elevation": "Élévation",
     "guideControls": "Commandes du guide",
-    "selectLayerAndScrollChart": "Voir dans le panneau Graphique",
+    "selectLayerAndScrollChart": "Accéder au panneau Graphique",
     "featureInfoTable": "Informations sur les éléments",
     "featureInfoTableField": "Champ",
     "featureInfoTableValue": "Valeur",
@@ -455,8 +455,8 @@
     "noFeatures": "inconnue...",
     "showing": "affiché(s)",
     "total": "total",
-    "selectLayerAndScroll": "Voir dans le panneau Données",
-    "accessAdvancedFunctions": "Accéder au panneau Données pour les fonctions avancées",
+    "selectLayerAndScroll": "Accéder au panneau Données",
+    "accessAdvancedFunctions": "Accéder au panneau Données",
     "filterToggle": "les filtres",
     "searchInputLabel": "Rechercher dans les colonnes du tableau",
     "tableControls": "Contrôles du tableau de données",
