+++ packages/geoview-core/public/templates/demos/demo-custom-footer-height.html
@@ -24,9 +24,9 @@
     <div class="demo-section">
       <h4 id="HMap1">Footer height: 400px</h4>
       <textarea id="configTextArea1" cols="100" rows="5" readonly="">
-&lt;div id="Map1" class="geoview-map" 
-  data-lang="en" 
-  data-footer-height="400px" 
+&lt;div id="Map1" class="geoview-map"
+  data-lang="en"
+  data-footer-height="400px"
   data-config-url="../configs/footer-tabs-config.json"&gt;&lt;/div&gt;
       </textarea>
     </div>
@@ -35,9 +35,9 @@
     <div class="demo-section">
       <h4 id="HMap2">Footer height: 100vh</h4>
       <textarea id="configTextArea2" cols="100" rows="5" readonly="">
-&lt;div id="Map2" class="geoview-map" 
-  data-lang="en" 
-  data-footer-height="100vh" 
+&lt;div id="Map2" class="geoview-map"
+  data-lang="en"
+  data-footer-height="100vh"
   data-config-url="../configs/footer-tabs-config.json"&gt;&lt;/div&gt;
       </textarea>
     </div>
